[SourceKit] Use serial queue to handle XPC messages
Some request kinds must be handled in order the SourceKit received.
Using concurrent queue for the message handling breaks that, and forces
clients wait for the response before sending the subsequent requests.

Use a serial queue to handle messages, and dispatch to concurrent queues
to handle requests concurrently only if possible.

This commit only changes the main message handling queue from
"concurrent" to "serial". The actual request handling logic is meant to
be updated in later commits.