utils: Support multiple CMake versions in build.ps1
Starting with CMake 3.30, CMake is passing linker flags to the Swift
invocation used as the linker driver. This causes issues because Swift
does not understand the linker flags and fails to link.

This is essentially a reland of #84306 with a few changes:
* A new helper cmdlet takes care of adding the "right" version of the
  linker flags depending on the CMake version and the project
  configuration.
* Android linker flags are handled differently to work around issues in
  the NDK CMake files. In addition, we need to pass a linker driver flag
  for C/C++/ASM libraries to use the just built ld.lld linker. This
  argument is not understood by the Swift driver so it can only be used
  with clang used as the linker driver.

This also fixes a minor path normalization issue in swift shims, which
causes a build failure with CMake 3.30+.